Updated List of Proposed Prototypes Iteration 2
Our team aims to create a web application for event creation and user check-in, featuring a secure database for storing User & Admin login information and event details. We will design a user-friendly interface to facilitate easy event browsing and registration. The application will provide admins with data and analytics for monitoring event success and send registered users an email with a QR code for check-in. The prototype includes basic functionalities such as event creation, registration, ticket generation, and cart features for users to add multiple events.